requirements.txt
jsonschema==2.5.1
pyyaml==3.12